% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardi_mortality.R
\docType{data}
\name{cardi_mortality}
\alias{cardi_mortality}
\title{Cardiovascular disease on the mortality of COVID-19}
\format{
An object of class \code{data.frame} with 6 rows and 3 columns.
}
\usage{
data("cardi_mortality")
}
\description{
Cardiovascular disease on the mortality of COVID-19
}
\examples{
data("cardi_mortality")
}
\keyword{datasets}
