% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardi_severe.R
\docType{data}
\name{cardi_severe}
\alias{cardi_severe}
\title{Cardiovascular disease on the severe case rate of COVID-19}
\format{
An object of class \code{data.frame} with 6 rows and 3 columns.
}
\usage{
data("cardi_severe")
}
\description{
Cardiovascular disease on the severe case rate of COVID-19
}
\examples{
data("cardi_severe")
}
\keyword{datasets}
