% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egger.R
\name{egger}
\alias{egger}
\title{Egger test statistics}
\usage{
egger(beta, se2, phi2, m)
}
\arguments{
\item{beta}{The original or the simulated estimated effects.}

\item{se2}{The squared standard errors of the estimated effects.}

\item{phi2}{The value of the hyperparameter phi.}

\item{m}{The number of replications}
}
\value{
The egger test statistic value

@export
}
\description{
Egger test statistics
}
