% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{prob_to_r}
\alias{prob_to_r}
\title{Probability and r parameter value 1-1 transformation}
\usage{
prob_to_r(p)
}
\arguments{
\item{p}{A value, the required probability of sign consistency.}
}
\value{
The corresponding heterogeneity parameter value.
}
\description{
This function transforms the probability of simulated \eqn {\beta_j} having the same
sign with the underlying true effect \eqn{ \bar\beta } to the corresponding heterogeneity
r parameter value.
}
