% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpp.R
\docType{data}
\name{rpp}
\alias{rpp}
\title{Filtered RPP data}
\format{
An object of class \code{data.frame} with 73 rows and 5 columns.
}
\usage{
data("rpp")
}
\description{
Filtered RPP data
}
\examples{
data("rpp")
}
\keyword{datasets}
