% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\docType{data}
\name{severity}
\alias{severity}
\title{Cardiovascular disease impact on the severe case rate of COVID-19}
\format{
An object of class \code{data.frame} with 6 rows and 3 columns.
}
\usage{
data("severity")
}
\description{
Cardiovascular disease impact on the severe case rate of COVID-19
}
\examples{
data("severity")
}
\keyword{datasets}
